need to remove all init_args
